+++ backend/canisters/group/api/src/updates/edit_message_v2.rs
@@ -1,6 +1,6 @@
 use serde::{Deserialize, Serialize};
 use ts_export::ts_export;
-use types::{MessageContentInitial, MessageId, MessageIndex, UnitResult};
+use types::{MessageContentInitial, MessageId, MessageIndex, OgPreview, UnitResult};
 
 #[ts_export(group, edit_message)]
 #[derive(Serialize, Deserialize, Debug)]
@@ -10,6 +10,8 @@ pub struct Args {
     pub content: MessageContentInitial,
     pub block_level_markdown: Option<bool>,
     pub new_achievement: bool,
+    #[serde(default)]
+    pub og_previews: Vec<OgPreview>,
 }
 
 pub type Response = UnitResult;