+++ backend/canisters/user/impl/src/updates/start_video_call.rs
@@ -102,6 +102,7 @@ pub fn handle_start_video_call(
         forwarded: false,
         sender_is_bot: true,
         block_level_markdown: false,
+        og_previews: Vec::new(),
         now,
         sender_context: None,
     };