+++ backend/integration_tests/src/edit_message_tests.rs
@@ -71,6 +71,7 @@ fn update_block_level_markdown_succeeds(starting_value: bool) {
             block_level_markdown: starting_value,
             message_filter_failed: None,
             pin: None,
+            og_previews: Vec::new(),
         },
     ) else {
         panic!()