+++ backend/integration_tests/src/mentions_tests.rs
@@ -50,6 +50,7 @@ fn mention_users_succeeds(mention_everyone: bool) {
             rules_accepted: None,
             message_filter_failed: None,
             new_achievement: false,
+            og_previews: Vec::new(),
         },
     );
 
@@ -127,6 +128,7 @@ fn mention_everyone_only_succeeds_if_authorized(authorized: bool) {
             rules_accepted: None,
             message_filter_failed: None,
             new_achievement: false,
+            og_previews: Vec::new(),
         },
     );
 
@@ -190,6 +192,7 @@ fn mentioned_in_thread_adds_user_as_follower() {
             rules_accepted: None,
             message_filter_failed: None,
             new_achievement: false,
+            og_previews: Vec::new(),
         },
     );