+++ frontend/openchat-client/src/openchat.ts
@@ -3803,6 +3803,8 @@ export class OpenChat {
             msg.blockLevelMarkdown,
             [],
             true,
+            undefined,
+            msg.ogPreviews,
         );
     }
 
@@ -4321,6 +4323,7 @@ export class OpenChat {
         mentioned: User[] = [],
         forwarded: boolean = false,
         messageIdIfRetrying?: bigint,
+        ogPreviews: Message["ogPreviews"] = [],
     ): Promise<SendMessageResponse> {
         const { chatId, threadRootMessageIndex } = messageContext;
         const chat = allChatsStore.value.get(chatId);
@@ -4339,6 +4342,7 @@ export class OpenChat {
             repliesTo: draftMessage?.replyingTo,
             forwarded,
             blockLevelMarkdown,
+            ogPreviews,
         };
 
         return this.#sendMessageCommon(
@@ -9384,6 +9388,7 @@ export class OpenChat {
                 forwarded: false,
                 blockLevelMarkdown,
                 senderContext: botContext,
+                ogPreviews: [],
             });
             localUpdates.addUnconfirmed(context, event);
             publish("sentMessage", { context, event });